Feature/pn 217 generazione atti opponibili a terzi (#7)
* aggiunta tipologia indirizzo digitale nel recipient nella generazione atti opponibili a terzi

* aggiunta generazione atti opponibili a terzi su azione di invio pec

* integrazione test generazione atti opponibili a terzi

* implementazione generazione atti opponibili a terzi raggruppato su azione invio notifica di cortesia

* Spostato invio messaggi di cortesia ad inizio del workflow e unificati legal facts di spedizione
