+++ src/main/java/it/pagopa/pn/deliverypush/actions/ChooseDeliveryModeActionHandler.java
@@ -1,11 +1,14 @@
 package it.pagopa.pn.deliverypush.actions;
 
+import it.pagopa.pn.api.dto.events.*;
 import it.pagopa.pn.api.dto.notification.Notification;
 import it.pagopa.pn.api.dto.notification.NotificationRecipient;
+import it.pagopa.pn.api.dto.notification.address.DigitalAddress;
 import it.pagopa.pn.api.dto.notification.timeline.DeliveryMode;
 import it.pagopa.pn.api.dto.notification.timeline.NotificationPathChooseDetails;
 import it.pagopa.pn.api.dto.notification.timeline.TimelineElement;
 import it.pagopa.pn.api.dto.notification.timeline.TimelineElementCategory;
+import it.pagopa.pn.commons.abstractions.MomProducer;
 import it.pagopa.pn.commons.pnclients.addressbook.AddressBook;
 import it.pagopa.pn.commons_delivery.middleware.TimelineDao;
 import it.pagopa.pn.deliverypush.PnDeliveryPushConfigs;
@@ -16,17 +19,22 @@ import it.pagopa.pn.deliverypush.abstractions.actionspool.DigitalAddressSource;
 import org.springframework.stereotype.Component;
 
 import java.time.Instant;
+import java.util.List;
 
 @Component
 public class ChooseDeliveryModeActionHandler extends AbstractActionHandler {
 
     private final AddressBook addressBook;
     private final PnDeliveryPushConfigs pnDeliveryPushConfigs;
+    private final MomProducer<PnExtChnEmailEvent> emailRequestProducer;
 
-    public ChooseDeliveryModeActionHandler(TimelineDao timelineDao, AddressBook addressBook, ActionsPool actionsPool, PnDeliveryPushConfigs pnDeliveryPushConfigs) {
+    public ChooseDeliveryModeActionHandler(TimelineDao timelineDao, AddressBook addressBook,
+                      ActionsPool actionsPool, PnDeliveryPushConfigs pnDeliveryPushConfigs,
+                                           MomProducer<PnExtChnEmailEvent> emailRequestProducer) {
         super( timelineDao, actionsPool , pnDeliveryPushConfigs);
         this.addressBook = addressBook;
         this.pnDeliveryPushConfigs = pnDeliveryPushConfigs;
+        this.emailRequestProducer = emailRequestProducer;
     }
 
     @Override
@@ -45,12 +53,40 @@ public class ChooseDeliveryModeActionHandler extends AbstractActionHandler {
                 ;
 
         addressBook.getAddresses( recipient.getTaxId() )
-            .ifPresent( abEntry ->
+            .ifPresent( abEntry -> {
                 timelineDetailsBuilder
-                        .general( abEntry.getDigitalAddresses().getGeneral() )
-                        .platform( abEntry.getDigitalAddresses().getPlatform() )
-                        .courtesyAddresses( abEntry.getCourtesyAddresses() )
-            );
+                        .general(abEntry.getDigitalAddresses().getGeneral())
+                        .platform(abEntry.getDigitalAddresses().getPlatform())
+                        .courtesyAddresses(abEntry.getCourtesyAddresses());
+
+                // - Send Email
+                List<DigitalAddress> courtesyAddresses = abEntry.getCourtesyAddresses();
+                if( courtesyAddresses != null ) {
+                    int numberOfAddresses = courtesyAddresses.size();
+
+                    for (int idx = 0; idx < numberOfAddresses; idx++) {
+                        DigitalAddress emailAddress = courtesyAddresses.get(idx);
+                        this.emailRequestProducer.push(PnExtChnEmailEvent.builder()
+                                .header(StandardEventHeader.builder()
+                                        .iun(action.getIun())
+                                        .eventId(action.getActionId() + "_" + idx)
+                                        .eventType(EventType.SEND_COURTESY_EMAIL.name())
+                                        .publisher(EventPublisher.DELIVERY_PUSH.name())
+                                        .createdAt(Instant.now())
+                                        .build()
+                                )
+                                .payload(PnExtChnEmailEventPayload.builder()
+                                        .iun(notification.getIun())
+                                        .senderId(notification.getSender().getPaId())
+                                        .emailAddress(emailAddress.getAddress())
+                                        .build()
+                                )
+                                .build()
+                        );
+                    }
+                }
+
+            });
 
         NotificationPathChooseDetails timelineDetails = timelineDetailsBuilder.build();
