+++ src/main/java/it/pagopa/pn/deliverypush/actions/ReceivePecActionHandler.java
@@ -1,35 +1,44 @@
 package it.pagopa.pn.deliverypush.actions;
 
+import java.util.Collections;
+import java.util.Optional;
+
+import org.springframework.stereotype.Component;
+
 import it.pagopa.pn.api.dto.events.PnExtChnProgressStatus;
 import it.pagopa.pn.api.dto.notification.Notification;
 import it.pagopa.pn.api.dto.notification.NotificationRecipient;
 import it.pagopa.pn.api.dto.notification.address.DigitalAddress;
-import it.pagopa.pn.api.dto.notification.timeline.*;
+import it.pagopa.pn.api.dto.notification.timeline.NotificationPathChooseDetails;
+import it.pagopa.pn.api.dto.notification.timeline.SendDigitalDetails;
+import it.pagopa.pn.api.dto.notification.timeline.SendDigitalFeedbackDetails;
+import it.pagopa.pn.api.dto.notification.timeline.TimelineElement;
+import it.pagopa.pn.api.dto.notification.timeline.TimelineElementCategory;
 import it.pagopa.pn.commons.exceptions.PnInternalException;
 import it.pagopa.pn.commons_delivery.middleware.TimelineDao;
 import it.pagopa.pn.deliverypush.PnDeliveryPushConfigs;
 import it.pagopa.pn.deliverypush.abstractions.actionspool.Action;
 import it.pagopa.pn.deliverypush.abstractions.actionspool.ActionType;
 import it.pagopa.pn.deliverypush.abstractions.actionspool.ActionsPool;
-import org.springframework.stereotype.Component;
-
-import java.util.Collections;
-import java.util.Optional;
 
 @Component
 public class ReceivePecActionHandler extends AbstractActionHandler {
 
-    public ReceivePecActionHandler(TimelineDao timelineDao, ActionsPool actionsPool , PnDeliveryPushConfigs pnDeliveryPushConfigs ) {
+	private final LegalFactUtils legalFactStore;
+	
+    public ReceivePecActionHandler(LegalFactUtils legalFactStore, TimelineDao timelineDao, ActionsPool actionsPool , PnDeliveryPushConfigs pnDeliveryPushConfigs ) {
         super( timelineDao, actionsPool , pnDeliveryPushConfigs);
+        this.legalFactStore = legalFactStore;
     }
 
     @Override
     public void handleAction(Action action, Notification notification ) {
 
         Action nextAction;
+        PnExtChnProgressStatus status = action.getResponseStatus();
+        NotificationRecipient recipient = notification.getRecipients().get( action.getRecipientIndex() );
 
         // - Se il messaggio è andato a buon fine schedula l'attesa
-        PnExtChnProgressStatus status = action.getResponseStatus();
         if ( PnExtChnProgressStatus.OK.equals( status ) ) {
             nextAction = buildSendCourtesyAction(action);
         }
@@ -41,13 +50,9 @@ public class ReceivePecActionHandler extends AbstractActionHandler {
         scheduleAction( nextAction );
 
 
-        NotificationRecipient recipient = notification.getRecipients().get(action.getRecipientIndex());
-
         Optional<NotificationPathChooseDetails> addresses =
                 getTimelineElement( action, ActionType.CHOOSE_DELIVERY_MODE, NotificationPathChooseDetails.class );
-
         if( addresses.isPresent() ) {
-
             // - send pec if specific address present
             DigitalAddress address = action.getDigitalAddressSource().getAddressFrom(addresses.get());
             addTimelineElement(action, TimelineElement.builder()
